split e2e tests
